requirements-dev.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
pytest
pytest-asyncio
httpx
pytest-order